[UPDATE] : refactor modules to gain better separation
