+++ src/grid.rs
@@ -1,40 +1,7 @@
 use rand::Rng;
 use std::fmt;
-use std::fs;
-use std::io::Write;
 
-mod cell {
-    /// The possible status that can have a minesweeper cell
-    #[derive(Copy, Clone, PartialEq, Debug)]
-    pub enum Status {
-        /// There is no mine on this cell or next to this one
-        Nothing,
-        /// There is a mine on this cell
-        Mine,
-        /// There is one or multiple mine next to this cell,
-        /// the number of close mine is associated to this Status
-        NearMine(u8),
-    }
-
-    /// A cell of the minesweeper grid
-    #[derive(Copy, Clone, PartialEq, Debug)]
-    pub struct Cell {
-        /// The status (or content) of a cell (is there is a mine, nothing, an hints?)
-        pub status: Status,
-        /// Does the cell have been discovered by the player
-        pub discovered: bool,
-    }
-
-    impl Cell {
-        /// Returns an empty cell which has not been discovered
-        pub fn new() -> Self {
-            Cell {
-                status: Status::Nothing,
-                discovered: false,
-            }
-        }
-    }
-}
+mod cell;
 
 /// The minesweeper grid
 #[derive(PartialEq, Debug)]
@@ -49,7 +16,7 @@ static STANDARD_DIMENSIONS: (usize, usize) = (10usize, 10usize);
 
 impl Grid {
     /// Returns a 10x10 grid filled of undiscovered and empty cells
-    pub fn make_empty() -> Self {
+    fn make_empty() -> Self {
         Grid {
             dimensions: STANDARD_DIMENSIONS,
             rows: vec![vec![cell::Cell::new(); STANDARD_DIMENSIONS.1]; STANDARD_DIMENSIONS.0],
@@ -70,13 +37,13 @@ impl Grid {
     }
 
     /// Check for a grid position the number of mines next to it.
-    /// 
+    ///
     /// # Arguments
-    /// 
+    ///
     /// * `position` - The center cell on which all next cells will be checked
-    /// 
+    ///
     /// # Examples
-    /// 
+    ///
     /// ```rust
     /// let debug_str = "\
     ///                 X**\
@@ -122,7 +89,7 @@ impl Grid {
     }
 
     /// Replace each cells which is not a mine and that have a mines next to it
-    /// with the status Status::NearMine(x), with x the number of mines around it 
+    /// with the status Status::NearMine(x), with x the number of mines around it
     fn generate_hints(&mut self) {
         for x in 0..self.dimensions.0 {
             for y in 0..self.dimensions.1 {
@@ -143,130 +110,6 @@ impl Grid {
         grid.generate_mines();
         grid.generate_hints();
         grid
-    }
-
-    /// Returns a grid following a string
-    /// 
-    /// # Arguments
-    /// 
-    /// * `str` - the str from which the grid will be made
-    /// 
-    /// # Examples
-    /// 
-    /// ```rust
-    /// let debug_str = "\
-    ///                 ***\
-    ///                 *X*\
-    ///                 ***\
-    ///             ";
-    /// let grid = Grid::from_debug_str(debug_str);
-    /// ```
-    pub fn from_debug_str(str: &str) -> Self {
-        let mut dimensions = (0usize, 0usize);
-        let mut rows: Vec<Vec<cell::Cell>> = Vec::new();
-        let mut row: Vec<cell::Cell> = Vec::new();
-
-        for c in str.chars() {
-            match c {
-                'X' => row.push(cell::Cell {
-                    status: cell::Status::Mine,
-                    discovered: false,
-                }),
-                '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8' => {
-                    let nb_mines = c.to_digit(10).unwrap() as u8;
-
-                    row.push(cell::Cell {
-                        status: cell::Status::NearMine(nb_mines),
-                        discovered: false,
-                    });
-                }
-                '\n' => {
-                    if row.len() > dimensions.1 {
-                        dimensions.1 = row.len();
-                    }
-                    if row.len() > 0 {
-                        rows.push(row.clone());
-                        row.clear();
-                    }
-                }
-                '*' | _ => row.push(cell::Cell {
-                    status: cell::Status::Nothing,
-                    discovered: false,
-                }),
-            }
-        }
-        if row.len() > 0 {
-            rows.push(row.clone());
-        }
-        dimensions.0 = rows.len();
-        Grid {
-            dimensions: dimensions,
-            rows: rows,
-        }
-    }
-
-    /// Returns a grid following a debug file
-    /// 
-    /// # Arguments
-    /// 
-    /// * `file_path` - The file path to the debug file
-    /// 
-    /// # Examples
-    /// 
-    /// ```rust
-    /// /*
-    ///     "debug_file" content:
-    ///     ***
-    ///     *X*
-    ///     ***
-    ///  */
-    /// let grid = Grid::from_debug_file("./path/to/debug_file");
-    /// ```
-    pub fn from_debug_file(file_path: &str) -> Self {
-        let contents = fs::read_to_string(file_path).expect("error: cannot open file");
-
-        Grid::from_debug_str(&contents)
-    }
-
-    /// Save the grid into a debug file
-    /// 
-    /// # Arguments
-    /// 
-    /// * `file_path` - The file path to save
-    /// 
-    /// # Examples
-    /// 
-    /// ```rust
-    /// let grid = Grid::make_empty();
-    /// 
-    /// grid.to_debug_file("path/to/debug_file");
-    /// ```
-    pub fn to_debug_file(&self, file_path: &str) {
-        let mut file = fs::File::create(file_path).unwrap();
-
-        writeln!(file, "{}", self.to_debug_str()).unwrap();
-    }
-
-    /// Returns the grid into the minesweeper debug string format
-    pub fn to_debug_str(&self) -> String {
-        let mut str = String::new();
-
-        for (x, row) in self.rows.iter().enumerate() {
-            for cell in row {
-                match cell.status {
-                    cell::Status::Nothing => str.push_str("*"),
-                    cell::Status::Mine => str.push_str("X"),
-                    cell::Status::NearMine(nb_mines) => match nb_mines {
-                        1..=9 => str.push_str(&nb_mines.to_string()),
-                        _ => str.push_str("?"),
-                    },
-                };
-            }
-            if !(x == (self.dimensions.0 - 1)) {
-                str.push_str("\n");
-            }
-        }
-        str
     }
 }
 
@@ -293,145 +136,254 @@ impl fmt::Display for Grid {
     }
 }
 
+#[cfg(test)]
 mod tests {
     use super::*;
+    use std::fs;
 
-    #[test]
-    fn test_from_debug_file_empty() {
-        let grid_ref = Grid::make_empty();
-        let grid_cmp = Grid::from_debug_file("./config/grid/debug/without_hints/empty");
+    impl Grid {
+        /// Returns a grid following a string
+        ///
+        /// # Arguments
+        ///
+        /// * `str` - the str from which the grid will be made
+        ///
+        /// # Examples
+        ///
+        /// ```rust
+        /// let debug_str = "\
+        ///                 ***\
+        ///                 *X*\
+        ///                 ***\
+        ///             ";
+        /// let grid = Grid::from_debug_str(debug_str);
+        /// ```
+        pub fn from_debug_str(str: &str) -> Self {
+            let mut dimensions = (0usize, 0usize);
+            let mut rows: Vec<Vec<cell::Cell>> = Vec::new();
+            let mut row: Vec<cell::Cell> = Vec::new();
 
-        assert_eq!(grid_ref, grid_cmp);
+            for c in str.chars() {
+                match c {
+                    'X' => row.push(cell::Cell {
+                        status: cell::Status::Mine,
+                        discovered: false,
+                    }),
+                    '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8' => {
+                        let nb_mines = c.to_digit(10).unwrap() as u8;
+
+                        row.push(cell::Cell {
+                            status: cell::Status::NearMine(nb_mines),
+                            discovered: false,
+                        });
+                    }
+                    '\n' => {
+                        if row.len() > dimensions.1 {
+                            dimensions.1 = row.len();
+                        }
+                        if row.len() > 0 {
+                            rows.push(row.clone());
+                            row.clear();
+                        }
+                    }
+                    '*' | _ => row.push(cell::Cell {
+                        status: cell::Status::Nothing,
+                        discovered: false,
+                    }),
+                }
+            }
+            if row.len() > 0 {
+                rows.push(row.clone());
+            }
+            dimensions.0 = rows.len();
+            Grid {
+                dimensions: dimensions,
+                rows: rows,
+            }
+        }
+
+        /// Returns a grid following a debug file
+        ///
+        /// # Arguments
+        ///
+        /// * `file_path` - The file path to the debug file
+        ///
+        /// # Examples
+        ///
+        /// ```rust
+        /// /*
+        ///     "debug_file" content:
+        ///     ***
+        ///     *X*
+        ///     ***
+        ///  */
+        /// let grid = Grid::from_debug_file("./path/to/debug_file");
+        /// ```
+        pub fn from_debug_file(file_path: &str) -> Self {
+            let contents = fs::read_to_string(file_path).expect("error: cannot open file");
+
+            Grid::from_debug_str(&contents)
+        }
+
+        /// Returns the grid into the minesweeper debug string format
+        pub fn to_debug_str(&self) -> String {
+            let mut str = String::new();
+
+            for (x, row) in self.rows.iter().enumerate() {
+                for cell in row {
+                    match cell.status {
+                        cell::Status::Nothing => str.push_str("*"),
+                        cell::Status::Mine => str.push_str("X"),
+                        cell::Status::NearMine(nb_mines) => match nb_mines {
+                            1..=9 => str.push_str(&nb_mines.to_string()),
+                            _ => str.push_str("?"),
+                        },
+                    };
+                }
+                if !(x == (self.dimensions.0 - 1)) {
+                    str.push_str("\n");
+                }
+            }
+            str
+        }
+    }
+    mod from_debug_file {
+        use super::*;
+
+        #[test]
+        fn empty() {
+            let grid_ref = Grid::make_empty();
+            let grid_cmp = Grid::from_debug_file("./config/grid/debug/without_hints/empty");
+
+            assert_eq!(grid_ref, grid_cmp);
+        }
+
+        #[test]
+        fn empty_with_final_break_line() {
+            let grid_ref = Grid::make_empty();
+            let grid_cmp = Grid::from_debug_file(
+                "./config/grid/debug/without_hints/empty_with_final_break_line",
+            );
+
+            assert_eq!(grid_ref, grid_cmp);
+        }
+
+        #[test]
+        fn empty_with_hole_lines() {
+            let grid_ref = Grid::make_empty();
+            let grid_cmp =
+                Grid::from_debug_file("./config/grid/debug/without_hints/empty_with_hole_lines");
+
+            assert_eq!(grid_ref, grid_cmp);
+        }
     }
 
-    #[test]
-    fn test_from_debug_file_empty_with_final_break_line() {
-        let grid_ref = Grid::make_empty();
-        let grid_cmp =
-            Grid::from_debug_file("./config/grid/debug/without_hints/empty_with_final_break_line");
+    mod from_debug_str {
+        use super::*;
 
-        assert_eq!(grid_ref, grid_cmp);
+        #[test]
+        fn empty() {
+            let grid_ref = Grid::from_debug_file("./config/grid/debug/without_hints/empty");
+            let debug_str = "\
+                                    **********\n\
+                                    **********\n\
+                                    **********\n\
+                                    **********\n\
+                                    **********\n\
+                                    **********\n\
+                                    **********\n\
+                                    **********\n\
+                                    **********\n\
+                                    **********\n\
+                                ";
+            let grid_cmp = Grid::from_debug_str(debug_str);
+
+            assert_eq!(grid_ref, grid_cmp);
+        }
+
+        #[test]
+        fn many_mines() {
+            let grid_ref = Grid::from_debug_file("./config/grid/debug/without_hints/many_mines");
+            let debug_str = "\
+                                    **********\n\
+                                    *X********\n\
+                                    **********\n\
+                                    *****X****\n\
+                                    ****X*X***\n\
+                                    *****X****\n\
+                                    ******X***\n\
+                                    **********\n\
+                                    X*********\n\
+                                    XXX*******\n\
+                                ";
+            let grid_cmp = Grid::from_debug_str(debug_str);
+
+            assert_eq!(grid_ref, grid_cmp);
+        }
+    }
+    mod count_near_mines {
+        use super::*;
+
+        // For each test, check number of mines next to positions 1,1
+
+        #[test]
+        fn one_mine() {
+            let debug_str = "\
+                                    *X*\n\
+                                    ***\n\
+                                    ***\n\
+                                ";
+            let grid = Grid::from_debug_str(debug_str);
+
+            let nb_mines = grid.count_near_mines((1, 1));
+            assert_eq!(1, nb_mines);
+        }
+
+        #[test]
+        fn many_mines() {
+            let debug_str = "\
+                                    X*X\n\
+                                    ***\n\
+                                    X*X\n\
+                                ";
+            let grid = Grid::from_debug_str(debug_str);
+
+            let nb_mines = grid.count_near_mines((1, 1));
+            assert_eq!(4, nb_mines);
+        }
     }
 
-    #[test]
-    fn test_from_debug_file_empty_with_hole_lines() {
-        let grid_ref = Grid::make_empty();
-        let grid_cmp =
-            Grid::from_debug_file("./config/grid/debug/without_hints/empty_with_hole_lines");
+    mod generate_hints {
+        use super::*;
 
-        assert_eq!(grid_ref, grid_cmp);
-    }
+        #[test]
+        fn one_mine() {
+            let grid_ref = Grid::from_debug_file("./config/grid/debug/with_hints/one_mine");
+            let mut grid_cmp = Grid::from_debug_file("./config/grid/debug/without_hints/one_mine");
 
-    #[test]
-    fn test_from_debug_str_empty() {
-        let grid_ref = Grid::from_debug_file("./config/grid/debug/without_hints/empty");
-        let debug_str = "\
-                                **********\n\
-                                **********\n\
-                                **********\n\
-                                **********\n\
-                                **********\n\
-                                **********\n\
-                                **********\n\
-                                **********\n\
-                                **********\n\
-                                **********\n\
-                            ";
-        let grid_cmp = Grid::from_debug_str(debug_str);
+            grid_cmp.generate_hints();
+            assert_eq!(grid_ref, grid_cmp);
+        }
 
-        assert_eq!(grid_ref, grid_cmp);
-    }
+        #[test]
+        fn corner_mines() {
+            let grid_ref = Grid::from_debug_file("./config/grid/debug/with_hints/corner_mines");
+            let mut grid_cmp =
+                Grid::from_debug_file("./config/grid/debug/without_hints/corner_mines");
 
-    #[test]
-    fn test_from_debug_str_many_mines() {
-        let grid_ref = Grid::from_debug_file("./config/grid/debug/without_hints/many_mines");
-        let debug_str = "\
-                                **********\n\
-                                *X********\n\
-                                **********\n\
-                                *****X****\n\
-                                ****X*X***\n\
-                                *****X****\n\
-                                ******X***\n\
-                                **********\n\
-                                X*********\n\
-                                XXX*******\n\
-                            ";
-        let grid_cmp = Grid::from_debug_str(debug_str);
+            grid_cmp.generate_hints();
+            assert_eq!(grid_ref, grid_cmp);
+        }
 
-        assert_eq!(grid_ref, grid_cmp);
-    }
+        #[test]
+        fn many_mines() {
+            let grid_ref = Grid::from_debug_file("./config/grid/debug/with_hints/many_mines");
+            let mut grid_cmp =
+                Grid::from_debug_file("./config/grid/debug/without_hints/many_mines");
 
-    #[test]
-    fn test_count_near_mines_one() {
-        // grid with mine in position 5,4
-        let debug_str = "\
-                                **********\n\
-                                **********\n\
-                                **********\n\
-                                **********\n\
-                                **********\n\
-                                ****X*****\n\
-                                **********\n\
-                                **********\n\
-                                **********\n\
-                                **********\n\
-                            ";
-        let grid = Grid::from_debug_str(debug_str);
-
-        let nb_mines = grid.count_near_mines((5, 5));
-        assert_eq!(1, nb_mines);
-    }
-
-    #[test]
-    fn test_count_near_mines_many() {
-        // grid with mines in position 6,4; 4,4; 6,6; 6, 5
-        let debug_str = "\
-                                **********\n\
-                                **********\n\
-                                **********\n\
-                                **********\n\
-                                ****X*****\n\
-                                **********\n\
-                                ****XXX***\n\
-                                **********\n\
-                                **********\n\
-                                **********\n\
-                            ";
-        let grid = Grid::from_debug_str(debug_str);
-
-        let nb_mines = grid.count_near_mines((5, 5));
-        assert_eq!(4, nb_mines);
-    }
-
-    #[test]
-    fn test_generates_hints_one_mine() {
-        let grid_ref = Grid::from_debug_file("./config/grid/debug/with_hints/one_mine");
-        let mut grid_cmp =
-            Grid::from_debug_file("./config/grid/debug/without_hints/one_mine");
-
-        grid_cmp.generate_hints();
-        assert_eq!(grid_ref, grid_cmp);
-    }
-
-    #[test]
-    fn test_generates_hints_corner_mines() {
-        let grid_ref = Grid::from_debug_file("./config/grid/debug/with_hints/corner_mines");
-        let mut grid_cmp =
-            Grid::from_debug_file("./config/grid/debug/without_hints/corner_mines");
-
-        grid_cmp.generate_hints();
-        assert_eq!(grid_ref, grid_cmp);
-    }
-
-    #[test]
-    fn test_generates_hints_many_mines() {
-        let grid_ref = Grid::from_debug_file("./config/grid/debug/with_hints/many_mines");
-        let mut grid_cmp =
-            Grid::from_debug_file("./config/grid/debug/without_hints/many_mines");
-
-        grid_cmp.generate_hints();
-        println!("{}", grid_cmp.to_debug_str());
-        assert_eq!(grid_ref, grid_cmp);
+            grid_cmp.generate_hints();
+            println!("{}", grid_cmp.to_debug_str());
+            assert_eq!(grid_ref, grid_cmp);
+        }
     }
 }
