+++ src/main.rs
@@ -1,7 +1,7 @@
 mod grid;
 
 fn main() {
-    let grid = grid::Grid::make_empty();
+    let grid = grid::Grid::make_random();
 
     println!("{}", grid);
 }
